// samples/react-modern-version-history-extension/src/extensions/betterVersionHistory/models/FieldTypes.ts
export enum FieldType {
    Integer = "Integer",
    Text = "Text",
    Note = "Note",
    DateTime = "DateTime",
    Counter = "Counter",
    Choice = "Choice",
    Lookup = "Lookup",
    LookupMulti = "LookupMulti",
    Boolean = "Boolean",
    Number = "Number",
    URL = "URL",
    Computed = "Computed",
    Guid = "Guid",
    MultiChoice = "MultiChoice",
    User = "User",
    UserMulti = "UserMulti",
    ContentTypeId = "ContentTypeId",
    Taxonomy = "TaxonomyFieldType",
    TaxonomyMulti = "TaxonomyFieldTypeMulti",
}

export enum DisplayFormat {
    DateOnly = 0,
    DateTime = 1,
}